+++ packages/sources/web-components/src/components/vtmn-text-input/vtmn-text-input.tsx
@@ -1,14 +1,14 @@
-import { Component, Prop, h } from '@stencil/core';
+import { Component, Prop, h, ComponentInterface } from '@stencil/core';
 
 @Component({
   tag: 'vtmn-text-input',
   styleUrl: 'vtmn-text-input.pcss',
 })
-export class VtmnTextInput {
+export class VtmnTextInput implements ComponentInterface {
   /**
    * The id of the text input.
    */
-  @Prop() id!: string;
+  @Prop() identifier!: string;
 
   /**
    * The label text of the text input.
@@ -47,7 +47,7 @@ export class VtmnTextInput {
 
   render() {
     return [
-      <label class="vtmn-text-input_label" htmlFor={this.id}>
+      <label class="vtmn-text-input_label" htmlFor={this.identifier}>
         {this.labelText}
       </label>,
 
@@ -59,7 +59,7 @@ export class VtmnTextInput {
           ]
             .filter(Boolean)
             .join(' ')}
-          id={this.id}
+          id={this.identifier}
           placeholder={this.placeholder}
           disabled={this.isDisabled}></textarea>
       ) : (
@@ -72,7 +72,7 @@ export class VtmnTextInput {
             ]
               .filter(Boolean)
               .join(' ')}
-            id={this.id}
+            id={this.identifier}
             placeholder={this.placeholder}
             disabled={this.isDisabled}
           />