+++ packages/sources/css/src/components/checkbox/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.2.3...@vtmn/css-checkbox@0.2.4) (2021-06-14)
+
+**Note:** Version bump only for package @vtmn/css-checkbox
+
+
+
+
+
+## [0.2.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.2.2...@vtmn/css-checkbox@0.2.3) (2021-06-14)
+
+**Note:** Version bump only for package @vtmn/css-checkbox
+
+
+
+
+
+## [0.2.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.2.1...@vtmn/css-checkbox@0.2.2) (2021-06-14)
+
+
+### Bug Fixes
+
+* **@vtmn/css-checkbox:** bug fixes + improvements ([#350](https://github.com/Decathlon/vitamin-web/issues/350)) ([bc1f1b2](https://github.com/Decathlon/vitamin-web/commit/bc1f1b2141573596e0468d5fcbe67daa6edae8d2))
+
+
+
+
+
+## [0.2.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.2.0...@vtmn/css-checkbox@0.2.1) (2021-06-14)
+
+**Note:** Version bump only for package @vtmn/css-checkbox
+
+
+
+
+
 # [0.2.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.1.2...@vtmn/css-checkbox@0.2.0) (2021-06-02)